requirements.txt
pygame
PyOpenGL
PyOpenGL_accelerate
numpy
glfw